[INTERNAL] removed some warnings
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/OpenMS@10036 6adb6e08-d915-0410-941f-83917bcadc18

+++ include/OpenMS/ANALYSIS/TARGETED/TargetedExperimentHelper.h
@@ -297,13 +297,13 @@ public:
         {
           if ((*peptide_group_labels_)[i] == label)
           {
-            peptide_group_label_ = i;
+            peptide_group_label_ = (Int)i;
             return;
           }
         }
 
         // not found, add it to the list
-        peptide_group_label_ = peptide_group_labels_->size();
+        peptide_group_label_ = (Int)peptide_group_labels_->size();
         peptide_group_labels_->push_back(label);
       }
